+++ packages/runtime/src/application/executor/merge-executor.ts
@@ -1,11 +1,10 @@
 import { resolveExpression } from "../../domain/dsl/expression";
 import {
   MergeExecutorError,
-  type MergeNodeDefinition,
-  type MergeStrategy,
   type RuntimeContext,
   type RuntimeStateStore
 } from "../../types";
+import type { MergeNodeDefinition, MergeStrategy } from "@zhongmiao/meta-lc-kernel";
 
 export type MergeExecutorResult = unknown;